requirements.txt
flask
requests
trafilatura
pypdf
sqlalchemy
python-dotenv
readability-lxml
beautifulsoup4
google-genai
